Move yq editor into the yaml package
Moving the yq editor into the yaml package so that it's more of a
generic editor for any yaml, not just the porter manifest. I'm going to
use it in our tests.
